moved update trade route
